sdm660-common: qcom-caf: audio: Fix compilation errors

+++ qcom-caf/audio/hal/audio_extn/sndmonitor.c
@@ -605,7 +605,7 @@ void *monitor_thread_loop(void *args __unused)
 // ---- listener static APIs ---- //
 static int hashfn(void *key)
 {
-    return (int)key;
+    return (intptr_t)key;
 }
 
 static bool hasheq(void *key1, void *key2)
